build: fix readthedocs dependency
Read the Docs was having a hard time with Python >=3.10 declared for this package. The error was actually thrown from Sphinx in the context of Read the Docs, but the same error coud not be reproduced locally. The error: module "types" has no attribute "UnionType". Reverting back to Python 3.9, a version used in another project and that is not encountering this error.