feat: ensure EML userId is a URL for linked data compatibility
Modify the EML userId element to be a URL, whenever possible,
facilitating linked data compatibility. If the current value isn't a URL,
it's converted using the directory attribute as a base URL.

This function addresses a practice, previously recommended by EDI, of
setting the base URL as the directory attribute and the remaining
identifier as the element value. This practice has been recently
deprecated.