refactor: pass EML to `annotate_workbook` subroutines for context
Add the EML path to the `annotate_workbook` function to enable context
for annotation subroutines like `add_dataset_annotations_to_workbook`.
This provides access to the original EML file, enabling the use of
descriptive text beyond the workbook's description field.

This is crucial for annotating elements like units, which require
information from the EML's raw unit name and which is not listed in the
workbook's description field.

This also enables annotators to source descriptive text directly from the
authoritative source (i.e. the EML). It's possible the workbook
descriptions could fall out of synch with the corresponding EML
document and this refactor addresses this.

Note, the description field provides context to human annotators, which
would otherwise have to browse data package landing pages, which is
arguably much less efficient. So we keep this information in the workbook
for now.